requirements.txt
anyio==3.6.2
black==22.12.0
click==8.1.3
colorama==0.4.6
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
